asr/test and presentation/requirements_deepspeech2_cpu.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
abseil-cpp=20200923.3=h9c3ff4c_0
absl-py=0.15.0=pyhd8ed1ab_0
aiohttp=3.7.4.post0=py38h497a2fe_0
appdirs=1.4.4=pyh9f0ad1d_0
astor=0.8.1=pyh9f0ad1d_0
astunparse=1.6.3=pyhd8ed1ab_0
async-timeout=3.0.1=py_1000
attrs=21.2.0=pyhd8ed1ab_0
audioread=2.1.9=py38h578d9bd_0
blinker=1.4=py_1
brotlipy=0.7.0=py38h497a2fe_1001
bzip2=1.0.8=h7f98852_4
c-ares=1.17.2=h7f98852_0
ca-certificates=2021.10.26=h06a4308_2
cachetools=4.2.4=pyhd8ed1ab_0
certifi=2021.10.8=py38h06a4308_0
cffi=1.14.6=py38h3931269_1
chardet=4.0.0=py38h578d9bd_1
click=8.0.3=py38h578d9bd_0
cryptography=35.0.0=py38ha5dfef3_0
cycler=0.10.0=py_2
dataclasses=0.8=pyhc8e2a94_3
decorator=5.1.0=pyhd8ed1ab_0
ffmpeg=4.4.0=h6987444_3
freetype=2.10.4=h0708190_1
gast=0.3.3=py_0
gettext=0.19.8.1=h73d1719_1008
giflib=5.2.1=h36c2ea0_2
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=1.35.0=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
google-pasta=0.2.0=pyh8c360ce_0
grpc-cpp=1.36.4=hf89561c_1
grpcio=1.36.1=py38hdd6454d_0
h5py=2.10.0=nompi_py38h9915d05_106
hdf5=1.10.6=nompi_h6a2412b_1114
icu=68.1=h58526e2_0
idna=2.10=pyh9f0ad1d_0
importlib-metadata=4.8.1=py38h578d9bd_0
jbig=2.1=h7f98852_2003
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9d=h36c2ea0_0
keras-applications=1.0.8=py_1
keras-base=2.3.1=py38_0
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.3.2=py38h1fd1430_0
krb5=1.19.2=hcc1bbae_2
lame=3.100=h7f98852_1001
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=3.0=h9c3ff4c_0
libblas=3.9.0=12_linux64_openblas
libcblas=3.9.0=12_linux64_openblas
libcurl=7.79.1=h2574ce0_1
libdeflate=1.8=h7f98852_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libffi=3.4.2=h9c3ff4c_4
libflac=1.3.3=h9c3ff4c_1
libgcc-ng=11.2.0=h1d223b6_11
libgfortran-ng=11.2.0=h69a702a_11
libgfortran5=11.2.0=h5c6108e_11
libgomp=11.2.0=h1d223b6_11
libiconv=1.16=h516909a_0
liblapack=3.9.0=12_linux64_openblas
libllvm11=11.1.0=hf817b99_2
libnghttp2=1.43.0=h812cca2_1
libnsl=2.0.0=h7f98852_0
libogg=1.3.4=h7f98852_1
libopenblas=0.3.18=pthreads_h8fe5266_0
libopus=1.3.1=h7f98852_1
libpng=1.6.37=h21135ba_2
libprotobuf=3.15.8=h780b84a_1
librosa=0.8.1=pyhd8ed1ab_0
libsndfile=1.0.31=h9c3ff4c_1
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=11.2.0=he4da1e4_11
libtiff=4.3.0=h6f004c6_2
libvorbis=1.3.7=h9c3ff4c_0
libvpx=1.11.0=h9c3ff4c_3
libwebp-base=1.2.1=h7f98852_0
libxml2=2.9.12=h72842e0_0
libzlib=1.2.11=h36c2ea0_1013
llvmlite=0.37.0=py38h4630a5e_0
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.3.4=pyhd8ed1ab_0
matplotlib-base=3.4.3=py38hf4fb855_1
multidict=5.2.0=py38h497a2fe_0
ncurses=6.2=h58526e2_4
nettle=3.6=he412f7d_0
numba=0.54.1=py38h4bf6c61_0
numpy=1.19.5=py38h9894fe3_2
oauthlib=3.1.1=pyhd8ed1ab_0
olefile=0.46=pyh9f0ad1d_1
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hb52868f_1
openssl=1.1.1l=h7f8727e_0
opt_einsum=3.3.0=pyhd8ed1ab_1
packaging=21.0=pyhd8ed1ab_0
pandas=1.3.4=py38h43a58ef_0
pillow=8.3.2=py38h8e6f84c_0
pip=21.3.1=pyhd8ed1ab_0
pooch=1.5.2=pyhd8ed1ab_0
protobuf=3.15.8=py38h709712a_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=pyh9f0ad1d_2
pyjwt=2.3.0=pyhd8ed1ab_0
pyopenssl=21.0.0=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pysocks=1.7.1=py38h578d9bd_3
pysoundfile=0.10.3.post1=pyhd3deb0d_0
python=3.8.12=hb7a2778_2_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-flatbuffers=1.12=pyhd8ed1ab_1
python_abi=3.8=2_cp38
pytz=2021.3=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=py38h497a2fe_0
re2=2021.04.01=h9c3ff4c_0
readline=8.1=h46c0cb4_0
requests=2.25.1=pyhd3deb0d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
resampy=0.2.2=py_0
rsa=4.7.2=pyh44b312d_0
scikit-learn=1.0=py38hacb3eff_1
scipy=1.7.1=py38h56a6a73_0
setuptools=58.2.0=py38h578d9bd_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.8=he1b5a44_3
sqlite=3.36.0=h9cd32fc_2
tensorboard=2.4.1=pyhd8ed1ab_1
tensorboard-plugin-wit=1.8.0=pyh44b312d_0
tensorflow-base=2.4.0=py38h01d9eeb_0
tensorflow-estimator=2.4.0=pyh9656e83_0
termcolor=1.1.0=py_2
threadpoolctl=3.0.0=pyh8a188c0_0
tk=8.6.11=h27826a3_1
tornado=6.1=py38h497a2fe_1
typing-extensions=3.10.0.2=hd8ed1ab_0
typing_extensions=3.10.0.2=pyha770c72_0
urllib3=1.26.7=pyhd8ed1ab_0
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.37.0=pyhd8ed1ab_1
wrapt=1.13.2=py38h497a2fe_0
x264=1!161.3030=h7f98852_1
xz=5.2.5=h516909a_1
yaml=0.2.5=h516909a_0
yarl=1.7.0=py38h497a2fe_0
zipp=3.6.0=pyhd8ed1ab_0
zlib=1.2.11=h36c2ea0_1013
zstd=1.5.0=ha95c52a_0
